Fix core/private layering with JsrtSourceHolder
Instead of implementing IActiveScriptByteCodeSource in JsrtByteCodeSourceMapper
and using DynamicSourceHolder, implement ISourceHolder itself.  It
eliminates one extra object and avoid core referencing code in private.
Also #if out the use of GetProcessInformation to get the available commit number
in core for now, as that requires Win10 SDK (Core still builds with Win8.1 SDK)