MSFT:4148269: Don't throw JavascriptError from ParseAsmJs
When parsing asm.js someone above us does EXIT_SCRIPT, so we should not
be throwing a JavascriptError on failed hresult from parser. Rather, we
should throw normal C++ error for OOM/OOS (which are the only errors
that can happen). These will be caught by the BytecodeGenerator and
then converted into the appropriate JavascriptError.