Fixes for FaultInjection bug in PageAllocator::Initialize and SetVirtualAllocator
Bug 4152207:
Fault Injection bug: We are not returning false (upon OOM) in
SegmentBase<T>::Initialize because we are checking with a wrong variable.

Bug 2863572 :
SetVirtualAllocator should be called only once, or if there are
multithreads - it can be called more than once but with the same value.
InterlockedExchangeCompare is used to lock the property.

Tests:
Exprgen, Crawler - Looks clean
UT - x64debug passed.
Perf looks flat.