Fix OS bug 4532699 lambda invalid parse
Parsing of an arrow function did not handle operator precedence quite
right.  It would allow arrow functions to appear on the right hand side
of operators such as logical or ||.  The => operator is
right-associative like assignment and is supposed to share the same
operator precedence as assignment, but we were skipping that.  This was
my bad from a couple years ago.

Fixed by handling the same way as assignment but ignoring all the
assignment special casing.  That is, fix by comparing opl < oplMin.

Code Reviewers: abchatra;pleath;tcare;tawoll