Have RegExp synchronize on short literals
Synchronizing on short literals improves performance when it reduces the
number of comparisons. One example is when the RegExp is /\ba\b/.
Previously, \b caused this RegExp to try to synchronize on a lot of
characters. However, with this chage, we first try to find the character
'a' and then test for \b.