Squashed commit of the following:
commit 5e7bd87b951e36d8196c9dfc4b967aa8fb08821d
Author: Rajat Dua <radua@microsoft.com>
Date:   Wed Sep 16 10:47:52 2015 -0700

    0916

commit fb370098570e3e609f00502bc26b06cac12341ec
Author: Rajat Dua <radua@microsoft.com>
Date:   Mon Sep 14 17:47:17 2015 -0700

    0914

Tail duplication can result in cases in which an outer loop lexically ends before the inner loop. The register allocator could then thrash in the inner loop registers used for a live-on-back-edge sym on the outer loop. To prevent this, we need to mark the end of the outer loop as the end of the inner loop and update the lifetimes already extended in the outer loop in keeping with this change.