Get core tests passing with ch.exe (and jshost)
1. Temporarily disable tests that are broken with ch pending fixes.
Disabled tests tagged with exclude_ch are blocked either by missing
or below parity functionality, or are written in a way that doesn't work
with ch.exe.

Note: AsmJs tests are all disabled for ch.exe.  They fail the vast
majority of tests on x86debug.  This includes the asmjs variants of the
SIMD tests.

2. Remove unnecessary -IgnoreScriptErrorCode switch from tests. It is not
supported by ch.exe and appears no longer necessary for jshost.

3. Fix up the manual stack trace path trimming used by the StackTrace
test suite to look for core\test --> This won't work when working just
in the core repo unless the user named it core!