Fixes OACR warning from turning on RunCodeAnalysis (/analyze aka prefast)
Except two warnings which I have notify owner to fix.

This also reduce the warning under core in Razzle from ~570 to 109.
The remaining warnings from razzle are from plugins not available in the
default VS2015 installation.

After all the warning is fixed, I will enable this in the CI and Dev build
on VSO, for the Core build.
However, it will still be off by default on the dev machine as the analyze
slow down build time by > 2x