Bug fix for loading windows.globalization.dll
Bug 3689045: JSFunFuzz:CAS::CheckinID:1423598:CHK JIT is causing a
bad code gen with switches:-maxinterpretcount:2
-bgjit- -printsystemexception -werexceptionsupport -bvt

We were hitting an assert that checks if windows.globalization.dll is
loaded or not. We are hitting this because of low memory of machine and
was got through Fault injection. However this is possible on normal
machines as well, so it is legitimate to not have windows.globalization.dll
loaded. We would further fail gracefully anyway and that's what happens on
fre builds. Having this assert prohibits fault injection from doing
further testing. So removing the assert.