Stop creating name properties on anonymous functions
The Bug: There were two issues one was causing getProperty and hasProperty to fail another
was just causing getProperty to fail. the First issue was the use of a
deffered type Handler that filtered for name properties. The result is we
would always return hasProperty true on anonymous functions. further when
the function was undeffered we would create empty string as the name.
The Second issue was the result of a spec change where
Function.prototype lost its name property between April and June 2015

The fix: The first fix was to create a deffered type handler just for
anonymous functions that did not filter the name property. The second fix
was to remove the name property creation on Function.prototype when we
call InitializeFunctionPrototype.