Fix bug 4415496: ReferenceError on strict mode eval('this;') when function called with null
The following situation produced a ReferenceError on _lexicalThisSlotSymbol:

    function f() {
        "use strict";
        eval('this;');
        eval('typeof (this);');
    }
    f.call(null);

The generated bytecode looked like this:

    000a   InitUndecl           R6
    000c   BrSrNeq_A            x:0017 (   6)  R2  R6
    0011   ScopedLdFld          R2 = R4.<_lexicalThisSlotSymbol>, R1 #0

The intent was to check whether R2 contained an Undecl, except that InitUndecl actually creates an Undecl object at a new address. Adding a new bytecode operator designed for comparing the value of a register against the Undecl value in the Library solves this problem.

Resulting bytecode looks like:

    000a   BrNotUndecl_A        x:0015 (   6)  R2
    000f   ScopedLdFld          R2 = R4.<_lexicalThisSlotSymbol>, R1 #0

This results in the correct behavior, which is to jump over the read of _lexicalThisSlotSymbol unless the register contains an Undecl value.

The change involved adding the BrNotUndecl_A opcode in the extended range and converting this opcode to BrNotAddr_A when it appears, in order to take advantage of the identical code path.

Thanks to curtism, pleath.