restrict signatures for add and compare ops in asm.js
We were a bit liberal on our validation of add and compare ops, which
could lead to functional issues in case of overflow (e.g. javascript
would not overflow, we will wrap). Note that this should only impact
asm.js code that is malformed. We had a few handwritten unit tests that
rely on our looser validation, which I have cleaned up.