Don't fail when a RegExp range is followed by a dash and another character
When we have a RegExp where a dash and a character follows a range where
the end of the range is in the Unicode supplementary plane, e.g.,
/\u{10000}-\u{10002}-\u{10001}/u, we treat the RegExp as if it contains
two ranges. If the third character's mathematical value is less than the
second one's, we throw a SyntaxError since such a backward going range
isn't allowed by the spec.

This change updates the RegExp verification to accept such a RegExp. The
underlying matcher already works correctly, so there is no need to
change that.