MSFT:5031609 - Assignment to a restricted property doesn't always throw
The restricted properties are 'caller' and 'arguments' and assignment to them should end up throwing most of the time.

This is implemented in the spec as such:

let thrower = () => { throw TypeError; };
Object.defineProperty(Function.prototype, 'caller', { set: thrower, get: thrower, configurable: true, writable: false });

Sloppy mode functions have own properties for 'caller' and 'arguments' which are set to null. These are not writable or configurable. We have this right.

Most other functions (strict mode, arrow, generators, classes and class methods, etc) do not have own restricted properties. When we access those properties, we flow up into Function.prototype and should    end up throwing. Our bug is that we do not return the thrower function as a setter accessor when we try to assign to a restricted property.

Also adds correct support for bound functions. They do not have restricted
own properties.

Fix issue with class ctor delete property. We were checking to see if the property was a function with a super reference and throwing an error there. This caused a property lookup whenever we delete from class constructor, which is wrong. Also removed the throw which is not in the spec.

Fix generator function instances getting caller/arguments properties in getOwnPropertyNames output. These are restricted in generator function instances.