Check the ignore overflow bit count to determine if a range always overflows after an imul. Added unittests to cover this case.
When doing an IMul, there is a case where we allow the int to overflow past 32 bits. In that case, the resulting int is unknown and the possible range is everything.
By fixing that, the code checking for possible bounds checks removal will not remove a bounds check using an integer that could have overflowed.