remove unsafe code for identifying templatized JIT functions in ETW
OS: Bug 5120262:
MS-CHAKRA-1-10 - Long Function Names Could Lead to a Write Outside of a Stack Buffer

This is an obvious security issue on functions > 256 characters in size.
The code was a temporary hack for investigating perf of templatized JIT,
and with templatized JIT disabled it never got removed.

Since templatized JIT is disabled and it's unclear that we would even
want to specifically isolate these functions in ETW in released product,
for now I'm commenting out the offending code (instead of removing since
it might be useful to have if we want to add back this functionality).