Fix casting in PeepsMD::PeepAssign.
In this context, we are trying to replace a 64-bit mov reg with a 32-bit mov to reg, which by default does zero extend on the top 32-bit.  So we are just treating the lower 32-bit as an unsigned value.
This would have no release build impact, only the assert that does the sign extend incorrect would trigger.
.