Use PushId in the core build version number and qfe on CI build
The pre-build script will query the tfs server for the push id and push date of the commit,
and generate a props file with the information.  The project then use it in the version
as the build number and qfe.  Also put the push id and push date in the description after the hash.

If it is a non-CI build, the buidl number and qfe are set to 65535.