Fix OS bug 4553357: async lambda parsing
`async => { }` did not parse with the async feature turned on.

The arrow function rewind and reparse was assuming that if the first
token was a tkID `async` then it was parsing an async arrow function,
but of course that is not always true as in the example above.

Fixed by rewinding again if `async` tkID is not followed by another tkID
nor a tkLParen.