Native unittest initial work
Pulled down sources of xunit++ framework (which includes .exe and .lib)
Create a test dll which will use the xunit++.lib. So that xunit++.console.exe can able to discover those test cases.
Ported on Jsrt Memory test a proof of concept and verified that it is portable and things work.