Native jsrt unittest initial work using xunit++
Pulled down sources of xunit++ framework (which includes sources .exe and .lib)
Create a test dll, jsrttests, which will use the xunit++.lib - using that xunit++.console.exe will be able to discover those test cases.
Ported one Jsrt Memory test as a proof of concept and verified that it is portable and things work.