Mark ES5ArrayIndexList.js test as slow
This test failed CI build and test of my rl timeout change because it
timed out, taking longer than the default 60 seconds.  I looked at the
test and determined it is not critical for normal dev testing so I have
marked it with the Slow tag to remove it from CI tests.