Refactored EmitCall to remove unused code paths and references to compat modes.
Each of the variants on EmitCallInstr and EmitCallTarget was only called from one location. For each, one of the variants was guarded by a flag fEvaluateComponents which was then passed into the function. This meant that a constant value was being passed into those function calls. I lowered the constant value of that field into the function and removed resulting dead code paths, and finally removed that now-unused parameter from the functions.

Finally, since the functions have clear uses in the modern code, I renamed them to remove the implication that they are compat versus modern usage, and I removed or edited comments to remove unnecessary references to compat mode.