Fix for the assert where we were wrongly converting the array/object to destructuring pattern.
When we realized that current grammar can be valid grammar for destructuring we didn't check that the current node is used for unary operator. Unary operators have higher precedence over assignment. Due to that we have wrongly converted the current node to destructuring node.
Fixed that by checking the oplMin (which tells that ParseExpr is looking for unary expression) is not greater than assignment.