Change for-in/of loops to have per iteration scope lifetime instead of per loop
Each iteration of a for-in or for-of loop with let or const declared
loop variables now has a new instance of those variables per iteration
of the loop as per the final ES6 spec.  This means captures of those
variables now properly capture each loop iteration's version of the
variable, rather than a single shared variable.

e.g.
```javascript
let a = [];
for (let x of [1, 2, 3]) {
    a.push(() => x);
}

for (let f of a) {
    print(a());
}
```

With this change we get the proper output:
```
1
2
3
```

instead of:
```
3
3
3
```