Rename ScopeSlots Scope methods to ScopeMetadata
I was getting confused about what the "Scope" on a ScopeSlots was since
the ScopeSlots itself is supposed to be the scope.  This "Scope" is
actually the metadata object used for debugging purposes, either the
FunctionBody of a function scope or a DebuggerScope object for an inner
block scope.

To make this clearer I have renamed the accessor methods to
SetScopeMetadata and GetScopeMetadataRaw.  Get is suffixed with Raw
because there are two other APIs for getting the metadata object cast to
either FunctionBody* or DebuggerScope* and are the primary use cases.
GetScopeMetadataRaw is only used once currently for copying the value to
from a StackScopeSlots to a heap ScopeSlots when boxing a stack
allocated scope slots.