Scope cloning for for-loop lexical bindings
Use the new bytecode ops CloneInnerScopeSlots and CloneBlockScope to
implement the ES6 per iteration bindings of lexical (let and const)
declared loop variables.

Change for-loop bytecode to include a scope clone after the initializer
expression and before the loop start, and then again after the body and
before the increment expression.