Handle CloneInnerScopeSlots and CloneBlockScope in the JIT backend.
In IRBuilder create the source operand but leave the destination empty.
Otherwise we will get boundary check failures later on in IRBuilder that
assume scope temps are single def.  We also do not want to use
BuildDstOpnd because it will create a new temp which is not what we
want. We want to use the same temp for the dst so we can make a simple
in place update of the scope slot array or activation object.

Assign the destination of the instruction in the FlowGraph instead.  Do
it here so that GlobOpt correctly kills copy prop on the scope temp so
that for example LdSlotArr is not hoisted outside the loop when cloning
scopes.