Fix up debugger to handle the for loop per iteration scoping
The debugger complained about the loop variable not having a bytecode
offset for where the variable is initialized with a real value.  However
in this particular scope the loop variable is never initialized and is
always in the TDZ.  To remedy this I added a method,
DebuggerScope::UpdatePropertiesInForInOrOfCollectionScope, that will
mark all the symbols in this debugger scope as belonging to this special
case.  The debugger then recognizes that these symbols are always in TDZ.