Fix Babylonjs es6 demo page script error "TypeError: Object doesn't support this action" thrown from class constructor
This script error is caused by inlined base class constructor receiving
a new object instead of a constructor function (aka "new.target").
The root cause is that logic was not implemented for fixed function
lowerer codepath.

This changeset adds the JIT-time check of constructor functionInfo for
(non-base-class) constructors and inserts the assignment of constructor
function.

Reviewed by: tawoll, pleath, abchatra