Drop support for manually assigned RL test names
We do a linear lookup for the name of each test and we do this as we
append new tests to the test list. If the number of tests is high, this
lengthens the RL setup time quite a lot.

Even though we support associating names with tests, we don't use this
feature in any of our test suites. Therefore, it would make sense to
remove this unused feature rather than changing the code to make the
lookup faster.