Stack Arguments Optimization extended - With Formal's presence.
Issue:

We optimize arguments object elements read by reading the values directly from the stack slot instead of creating a heap arguments object(as done in general case).
But this optimization works only in the absence of formals. We have few hot functions in node as reported by Kunal, that has arguments object and formals.

Fix:
This change extends the stack arguments optimization even when formals are
present. But we give up when there are writes to formals, as mapping will
be wrong.
Also, formals' accesses are still from the Activation object and hence is not optimized for this iteration and is the next workitem.

Test:
Exprgen looks fine. I have started other runs.

Perf:
Console perf runs look fine.