Quick implementation of section and module headers
This is an implementation of the string-form of section headers
and the new module header. It currently uses a suboptimal
chain of string comparisons, but this is a placeholder for
something more efficient as the discussion on whether to
use an LEB128-like encoding or brotli instead of general strings
is settled. Ideally, we would just hash the string using
JsUtil::CharacterBuffer<char>::StaticGetHashCode() or some faster
alternative and do lookup of the hash map. This commit will be
helpful to establish the section handling in preparation for
breaking up the function section into decls and defs.