Cut dependency on ScriptContext for Scanner and Parser
This commit removes the dependency that the Scanner and Parser have
on Js::ScriptContext. Instead, they take a dependency on a
ScriptContextParseFacade- this is a view over the ScriptContext
that is specific to the Scanner and Parser and contains all
data and method declarations that are needed for the parse phase.
The facade is templatized to allow mocking up of the ScriptContext
in the facade itself. This implies that the parser and scanner are
also templatized- the template type is the type of the implementation
of ScriptContext that is used by the facade.

This change doesn't build yet since the individual members of Scanner
and Parser still reference fields that the facade does not yet provide.
The next commit will fill in these fields.