MSFT: 7429352 JSRT context leak regression fix
We start supporting cross-context scenarios in JSRT in chakracore. However we are having
leaks in different scenarios due to pinned object. This change cleans up some basic leak scenarios.
We might need to have more cleanup in the future to make sure that there is only one
root for a JSRTContext or script context.

One major leak source is the guest arena in ScriptContext. objects allocated in the guestArena are
hard pinned till the scriptcontext is closed. We need to move objects off the guestArena:
    scriptContext::Cache. it's allocated using normal RecyclerNew, linked in the JavascriptLibrar, and
keep the reference in scriptcontext to minimize the code change.
    BuiltInFunctionMap. We might need to have a full cleanup of BindReference, but for now we just move
out the builtInFunctionMap to JavascripLibrary.
    Allocate JSRTContext using RecyclerNew (no more leaf); have JSRTContext have a different reference to
JavascriptLibrary to keep it alive, and masked out other fields for potential false reference.