ImplicitFlag_Accessor
 = o.x
p.a =    // setter
 = o.y

From <https://github.com/Microsoft/ChakraCore/pull/982>
For this example, if during interpretation we know that p.a was setter, we record that profile info and set implictflag to ImplicitFlag_Accessor for that instruction. For this case, we assumed that setter would get inlined and hence there is no point in tracking ImplicitFlag_Accessor flag. So we always ignored this flag inside ImplicitFlagAllowOpts(). However in this case, setter was not getting inlined and every time we called, we bailed out. In bailout, we would just check that the implicitFlag was set to NONE (which in this case was set to ImplicitFlag_Accessor) and we would not REJIT and we keep bailing out for the function. With my change, in bailout we would see if that BailoutReason was BailoutOnImplicitCall and if old and new flags were not changed and they were ImplicitFlag_Accessor, we would rejit and record in profile info to stop ignoring ImplicitFlag_Accessor for that function for next JIT. In ImplicitFlagAllowOpts() we would stop ignoring this flag.