Throw SyntaxError if arguments symbol is referenced in the param scope of a split scoped function
This changelist makes it a SyntaxError to use arguments symbol in the param scope in a split scoped function. Add a duplicate symbol in the body to allow the usage in the body. During BeginBodyScope we copy over the value of arguments symbol from the param scope to the corresponding symbol in the body.