Fixing an issue with special scope slots allocation in split scope
In split scope the scope slots for this, super, super constructor and
new.target are allocated in the param scope and are not visible in the
body scope. This changelist is to allocate a duplicate scope slot in the
body scope for these symbols and and copy over the value just after BeginBodyScope.