Made assert more strict
Earlier I was verifying if the object whose prototype is being changed
is not registered as prototype object of proto or accessor inline cache.
@curtisman suggested to do verification more stricter. So now updated
to verify that the object is not in prototype chain of any type registered
in inline cache. With that I do not need previous verifications that I added.