Adding typeof support to enable Stack arg optimization
Premise:

acme-benchmark uses arguments[i] with typeof keyword. We do not currently
optimize for this scenario.

Change:

Added support in the Globopt to enable Stack arg optimization when the
function body has typeof keyword.

Details:

We optimistically insert (BailOnStackArgsOutOfActualsRange) for
TypeOfElem, if it is a candidate for optimization.
We decide during DeadStore phase, whether this bail out is required or not
and then change the instr accordingly(as shown below).

After GlobOpt Phase:
Before:
dst = TypeOfElem arguments[i]

With this pack:
tmpdst = LdElemI_A arguments[i] <(BailOnStackArgsOutOfActualsRange)>
dst = TypeOf tmpdst