Fixing the arguments symbol behavior in split scope functions
For a split scope function currently we duplicate the arguments symbol in
the body and throw a SyntaxError if arguments is used in the param scope.
This fix removes that SyntaxError and fixes the scenario to have only one
symbol for arguments in the function. When arguments is referred to in the
body it is loaded from the param scope through param scope specific
bytecode.