Removing the condition to force scope slot for split scope formals
Right now in split scope case all the formals are forced to be put into
slots even though they are not captured. This change removes that work
around. This causes some changes in the debugger as previously we were
expecting all symbols to be in slots. When emitting locals in the body the
RegSlotWalker has to skip the symbols which are in the formal scope.