Fix incorrect type used for Async Spawn
AsyncSpawn bytecode did a hard cast of its function argument to
JavascriptGenerator* which was incorrect.  The type is
JavascriptGeneratorFunction*.  This turned out to be benign because it
was only used for a GetEntryPoint() call, which did the right thing
since it is part of RecyclableObject.

I changed the flow of that pointer through the code base to use the
right type and used ::FromVar to do the assert checked cast.

I also added knopAsyncSpawn to the -trace:parse dumper.