Change OP_NewScGenFunc to also do async functions
Introduce JavascriptAsyncFunction which is is a variant of
JavascriptGeneratorFunction.  It's only difference is the initial entry
point, which handles the internal (virtual) generator script function
differently than a normal generator.

The normal generator function entry point simply creates and returns a
generator object.  The async function entry point does "async spawn",
immediately calling into the virtual generator and setting it up to pump
with promises.  It returns the promise that wraps the remaining pending
work.