bugfix: setting appropriate arrayFlags
In my previous PR #1304, I changed the logic to set the arrayFlags to `ObjectArrayFlagsTag`
or `InitialArrayValue` depending on if size of Array passed to `Array` object ctor is 0 or not.
However this triggers `BailoutOnMissingValue` on first `SetItem` resulting in regression.
Inspecting interpreter code, realized that we should always set the arrayFlags for arrays
created using `Array` ctor to `InitialArrayValue` and set head segment length to 0. Fixed
this for `new Array(const)` case as well as `new Array(var)` case.